pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.71.0",
    "mcp>=1.19.0",
    "pyyaml>=6.0.3",
]
